+++ lib/models/charity.dart
@@ -1,0 +1,15 @@
+class Charity {
+  String name;
+  String id;
+  String description;
+  String website;
+  String paymentInfo;
+
+  Charity(String name, String id, String description, String website, String paymentInfo) {
+    this.name = name;
+    this.id = id;
+    this.description = description;
+    this.website = website;
+    this.paymentInfo = paymentInfo;
+  }
+}