+++ lib/models/player.dart
@@ -1,0 +1,17 @@
+class Player {
+  String id;
+  String name;
+  String age;
+  String position;
+  String team;
+  String league;
+
+  Player(String id, String name, String age, String position, String team, String league) {
+    this.id = id;
+    this.name = name;
+    this.age = age;
+    this.position = position;
+    this.team = team;
+    this.league = league;
+  }
+}